requirements.txt
pyyaml
elasticsearch>=7.0.0,<8.0.0